# Truncate data on reset / re-use existing metadata instance