After analyzing some slow-running unit test runs across a few services,
it became apparent that this could be largely attributed to
creating/dropping the database in between each test case.

Namely, there is a large cost attributed with evaluating each
table/column from scratch between tests. Locally, this could table up to
half a second(!). On some repos with 500+ tests, this adds up.

This change should reduce the cost of resetting the database (after the
initial run) milliseconds.

Tested this out on a number of repos, with substantial differences:
* 873 tests: 490s => 145s
* 422 tests: 131s => 30s

To see the difference in performance, see this flame graph comparison:
Before:
<img width="904" alt="Screen Shot 2019-12-09 at 3 57 22 PM" src="https://user-images.githubusercontent.com/19560680/70483204-a1ccdf00-1a9c-11ea-83d4-4d04d4a66638.png">

After:
<img width="915" alt="Screen Shot 2019-12-09 at 3 57 30 PM" src="https://user-images.githubusercontent.com/19560680/70483215-aa251a00-1a9c-11ea-94af-3b94c95aa30e.png">

Notice how many times `recreate_all` shows up in the "before" graph (it actually shows up more than once), vastly overshadowing the actual code being tested.

Diving into why it cost so much, you can see how much of the time was actually spent in `visit_metadata`, which happens because we were essentially re-calculating this data from scratch.

In "after", `recreate_all` actually *still* dominates a bunch of operations, but its relative proportion is much lower than before. Again, digging into that same function call afterwards, you can see that it goes right into SQL code execution, rather than metadata creation.